% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareAnnotations.R
\name{prepareREannotation}
\alias{prepareREannotation}
\title{Prepare repeat element annotations}
\usage{
prepareREannotation(rmsk)
}
\arguments{
\item{rmsk}{The RepeatMasker GRanges object}
}
\value{
An DataFrame object
}
\description{
Create unique annotation table for repeat element to their class and
families.
}
\examples{
rmsk <- readRDS(system.file('extdata', 'danRer11_rmsk_chr25_sub.RDS',
                             package='ratte'))
prepareREannotation(rmsk)
}
