% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareAnnotations.R
\name{prepareSalmonIndex}
\alias{prepareSalmonIndex}
\title{Create the index files for salmon}
\usage{
prepareSalmonIndex(
  salmonPath = "salmon",
  cpus = 2,
  FaFile = "rmsk.fa.gz",
  index = "TEindex",
  dryrun = TRUE,
  additionalParameters = "-n",
  ...
)
}
\arguments{
\item{salmonPath}{The path to salmon executable file.}

\item{cpus}{Threads number.}

\item{FaFile}{The fasta file.}

\item{index}{Output index folder.}

\item{dryrun}{If true, print the command only.}

\item{additionalParameters}{Additional parameters for salmon index.
By default '-n' will close the poly-A clip. If you did not remove
the poly-A in \link{prepareSeq} function, please remove '-n' parameter.}

\item{...}{Parameters passed to system2.}
}
\value{
If dryrun is false, the output of system2.
}
\description{
Run salmon index command for the given fasta.
}
\examples{
prepareSalmonIndex('salmon', dryrun=TRUE)

}
