% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollup.R
\name{rollup_RE}
\alias{rollup_RE}
\title{Rollup multiple salmon quants}
\usage{
rollup_RE(quant_sfs, qnames, seq_anno, select_feature, cpus = 2, retry = 3)
}
\arguments{
\item{quant_sfs}{The salmon quant.sf files path.}

\item{qnames}{Human-readable names for each quant.sf file.}

\item{seq_anno}{Output of \link{prepareSeq} function.}

\item{select_feature}{Type of feature. Available choices are:
'all', 'exon', 'intron', 'gene', or 'intergenic'.}

\item{cpus}{Number of threads.}

\item{retry}{Retry number for the rollup.}
}
\description{
Rollup the TPM and NumReads for repeat elements from salmon quants.
}
\examples{
quant_sfs <- dir(system.file('extdata', 'RNA-seq', package='ratte'),
        'quant.sf',
 full.names=TRUE)
qnames <- sub("^.*\\\\/(.*?).quant.sf", "\\\\1", quant_sfs)
rmsk <- readRDS(system.file('extdata', 'danRer11_rmsk_chr25_sub.RDS',
                             package='ratte'))

library(BSgenome.Drerio.UCSC.danRer11) # for sequences
library(TxDb.Drerio.UCSC.danRer11.refGene) # for genomic annotation
seq_anno <- prepareSeq(Drerio,
           TxDb.Drerio.UCSC.danRer11.refGene,
           rmsk,
           subsetGRanges=GRanges('chr25:11000000-15000000'))
re <- rollup_RE(quant_sfs, qnames, seq_anno, 'all')
re_gene_level <- rollup_RE(quant_sfs, qnames, seq_anno, 'gene')
}
