% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareAnnotations.R
\name{saveFasta}
\alias{saveFasta}
\title{Export the sequences as fasta file.}
\usage{
saveFasta(seq, filepath = "rmsk.fa.gz")
}
\arguments{
\item{seq}{The output of \link{prepareSeq} function.}

\item{filepath}{The output fasta file path.}
}
\value{
Logical value to indicate the success of output.
}
\description{
Export the sequences as a gzip fasta file.
}
\examples{
# library(AnnotationHub)
# ah <- AnnotationHub()
# (rmsk <- query(ah, c('RepeatMasker', 'Zebrafish')))
# rmsk <- rmsk[['AH98980']]
rmsk <- readRDS(system.file('extdata', 'danRer11_rmsk_chr25_sub.RDS',
                             package='ratte'))
library(BSgenome.Drerio.UCSC.danRer11) # for sequences
library(TxDb.Drerio.UCSC.danRer11.refGene) # for genomic annotation
seq <- prepareSeq(Drerio,
           TxDb.Drerio.UCSC.danRer11.refGene,
           rmsk,
           subsetGRanges=GRanges('chr25:11000000-15000000'))
outfile <- tempfile(fileext='.fa.gz')
saveFasta(seq, outfile)
}
